\name{@PDINFONAME@}
\alias{@PDINFONAME@}
\title{pdInfoBuilder annotation data package}
\description{

  Welcome to the @PDINFONAME@ annotation Package.  This is an
  microarray chipset specific package.  The purpose is to provide detailed
  information about the microarray and organism @ORGANISM@.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:@PDINFONAME@")} 
  
}
\examples{
  ls("package:@PDINFONAME@")
}
\keyword{datasets}
